requirements.txt
homeassistant==2025.1.1
pip>=21.3.1
ruff==0.8.6
pyserial-asyncio-fast==0.14
